third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/popup-unsafe-none-with-same-origin.https-expected.txt

This is a testharness.js-based test.
PASS unsafe-none document opening popup to https://web-platform.test:8444 with COOP: ""
PASS unsafe-none document opening popup to https://web-platform.test:8444 with COOP: "unsafe-none"
PASS unsafe-none document opening popup to https://web-platform.test:8444 with COOP: "jibberish"
PASS unsafe-none document opening popup to https://web-platform.test:8444 with COOP: "same-site"
PASS unsafe-none document opening popup to https://web-platform.test:8444 with COOP: "same-site unsafe-allow-outgoing"
PASS unsafe-none document opening popup to https://web-platform.test:8444 with COOP: "same-origin unsafe-allow-outgoing"
FAIL unsafe-none document opening popup to https://web-platform.test:8444 with COOP: "same-origin" assert_equals: Openee browsing context closed expected true but got false
FAIL unsafe-none document opening popup to https://web-platform.test:8444 with COOP: "same-origin-allow-popups" assert_equals: Openee browsing context closed expected true but got false
Harness: the test ran to completion.
